Custom Section Header creating constraint warnings
I am trying to create a dynamic section header for my UITableView.
However, I am getting all these nasty constraint warning when the
section header loads.

My custom section header is loaded from a custom Xib file that has four
rows of colour. Three of which has height set to 80 and the other one
has an aspect ratio of 1:1. Basically, I want to always display a
square UIView hence the 1:1 ratio.

Upon launch, I get what I expected, three rows of height 80 colours and
one 1:1 ratio colour UIView that spans across the screen. In the case
of Iphone 6, it is 394. However I get all these constraints warnings.

I also played with the estimatedSectionHeaderHeight and set it to [394
+ 80 * 3] and I still get the warnings.

Weird thing is that when I change estimatedSectionHeaderHeight to
something like [394 + 80 * 3 + 50], it appears that the section header
grows over the "section header frame" and starts to cover the section
row.

Basically, I'd like to have dynamic section header that can be of
random height based on the content size.